requirements.txt
flask
gunicorn
google-generativeai>=0.8.3
python-dotenv